Remove malformed context test case as correct behavior is not defined in the Standard
The FDC3 Standard does not define the expected behavior on receipt of a malformed context (by fdc3.open, or indeed any other function). Hence, it is not possible to adequately define test cases for it. There currently only exists a test case for fdc3.open which this PR will remove.