Only trigger NPM releases on release/* branch merges
THis PR remove the trigger for NPM releases on merges to the main branch, meaning that all NPM release going forward will come from release/* branches. This is a more explicit workflow which should be safer and makes it easier to do maintenance releases on past FDC3 versions when errors in them are identified.