replacing all schema links with pathname:// links so that they will display in future website previews
We've used both https://fdc3.finos.org/schemas links all over the place (these break when working on new schemas that are not yet published at the URL given) and relative links to files in the /static folder (you just give the path without the /static). The latter break in docusaurus 3 due to SPA routing - the correct URL is generated but docusaurus doesn't know how to display it and displays the 404 page - if you refresh the page so you are visiting it directly it works). By prefixing the realtive links with pathname:// they are opened as external links and work first time. They also work in previews. Hence this commit switches all schema links to the pathname:// format.